+++ src/main/java/com/gitsh01/libertyvillagers/cmds/VillagerInfo.java
@@ -112,6 +112,9 @@ public class VillagerInfo {
     public static List<Text> getEntityInfo(ServerWorld serverWorld, Entity entity) {
         List<Text> lines = new ArrayList<>();
         lines.add(new TranslatableText("text.LibertyVillagers.villagerInfo.title"));
+        if (entity == null) {
+            return lines;
+        }
         Text name = entity.getDisplayName();
         lines.add(new TranslatableText("text.LibertyVillagers.villagerInfo.name", name));