+++ src/main/java/com/gitsh01/libertyvillagers/config/GolemsConfig.java
@@ -39,4 +39,19 @@ public class GolemsConfig implements ConfigData {
 
     @ConfigEntry.Gui.Tooltip(count = 2)
     public int golemSpawnLimitRange = 128;
+
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public boolean golemStayNearBell = true;
+
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public int golemMaxBellRange = 128;
+
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public boolean golemMoveToShore = true;
+
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public int golemMoveToShoreRange = 10;
+
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public boolean golemsDontClimb = true;
 }